added default values for optional configs
